requirements.txt
pip>=20.1
Cython>=0.29.19
numpy>=1.17.4
matplotlib>=3.1.2
jupyter>=1.0.0
opencv-python>=3.4.9.31
multiprocess>=0.70.9
pandas>=0.25.3
seaborn>=0.9.0
torch==1.4.0
torchvision==0.5.0
fastai>=1.0.59<2.0.0
pillow>=6.2.1
pydub>=0.23.1
scipy>=1.4.1
pretrainedmodels>=0.7.4
xgboost>=0.90
geffnet>=0.9.5
eli5>=0.10.1
lightgbm>=2.3.1
scikit-learn>=0.21.3
albumentations>=0.4.3
tifffile
imagecodecs>=2020.5.30
maskslic>=0.3.3
imagehash>=4.1.0